+++ test_suite/tests/default.rs
@@ -7,14 +7,27 @@ use std::env;
 
 #[derive(Envconfig)]
 pub struct Config {
-    #[envconfig(from = "PORT")]
+    #[envconfig(from = "PORT", default = "5432")]
     pub port: u16,
 }
 
-#[test]
-fn test_when_default_is_not_specified_value_is_loaded_from_env_var() {
-    env::set_var("PORT", "8080");
+fn setup() {
+    env::remove_var("PORT");
+}
 
-    // let config = Config::init().unwrap();
-    // assert_eq!(config.port, 8080);
+#[test]
+fn test_when_env_is_missing() {
+    setup();
+
+    let config = Config::init().unwrap();
+    assert_eq!(config.port, 5432);
+}
+
+#[test]
+fn test_when_env_is_present() {
+    setup();
+
+    env::set_var("PORT", "8080");
+    let config = Config::init().unwrap();
+    assert_eq!(config.port, 8080);
 }